+++ src/collective/iconifiedcategory/__init__.py
@@ -11,5 +11,6 @@ logger = logging.getLogger('collective.iconifiedcategory')
 
 CAT_SEPARATOR = '_-_'
 CSS_SEPARATOR = '-'
+DEFAULT_FILESIZE_LIMIT = 5000000
 
 _ = MessageFactory('collective.iconifiedcategory')